Fix WP_HTML_Tag_Processor position when next_tag('button') fails in submit wrapper
Use bookmark/seek pattern to restore processor position when searching for
an inner button inside a non-button submit wrapper element. Without this,
a failed next_tag() call advances past remaining HTML, breaking the outer
traversal loop.

Also extract gutenblock render tests into dedicated Gutenblock_Render_Test file.
